refactor: split 3 oversized backend services into focused modules
- Extract enrichment_pipeline.py from sync/orchestrator.py (enrichment
  functions: _find_best_match, enrich_request_metadata, ISRC/URL helpers)
- Extract bridge_integration.py from now_playing.py (bridge functions:
  handle_now_playing_update, update_bridge_status, clear_now_playing)
- Extract play_history_service.py from now_playing.py (history functions:
  get_play_history, archive_to_history, get_next_play_order)
- Extract recommendation/deduplication.py from recommendation/service.py
  (3 dedup functions)
- Extract recommendation/query_builder.py from recommendation/service.py
  (Beatport/Tidal query builders)

All original modules re-export extracted functions for backward compat.
Test patch targets updated for enrichment_pipeline module path.
